pipelined batch/commit for sequential pipelines
